# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Alchemy API
  version: 2.0.2
  description: |
    API contract for Alchemy — an AI-driven recipe creation and management platform.
    Covers the user-facing v1 API at api.cookwithalchemy.com and the admin API at admin.cookwithalchemy.com.
servers:
  - url: https://api.cookwithalchemy.com/v1
security:
  - bearerAuth: []
paths:
...
  /metadata-jobs/retry:
    post:
      summary: Re-queue a metadata job safely (internal authenticated scope)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id]
              properties:
                job_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Retry status
          content:
            application/json:
              schema:
                type: object
                required: [ok, job]
                properties:
                  ok:
                    type: boolean
                  job:
                    type: object
                    required: [id, status, attempts, next_attempt_at]
                    properties:
                      id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [pending, processing, ready, failed]
                      attempts:
                        type: integer
                      next_attempt_at:
                        type: string
                        format: date-time
        default:
          $ref: '#/components/responses/ErrorResponse'
...
  /memory-jobs/retry:
    post:
      summary: Re-queue a memory job safely (internal authenticated scope)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [job_id]
              properties:
                job_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Retry status
          content:
            application/json:
              schema:
                type: object
                required: [ok, job]
                properties:
                  ok:
                    type: boolean
                  job:
                    type: object
                    required: [id, status, attempts, next_attempt_at]
                    properties:
                      id:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [pending, processing, ready, failed]
                      attempts:
                        type: integer
                      next_attempt_at:
                        type: string
                        format: date-time
        default:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ErrorResponse:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
        request_id:
          type: string
...